new version of rent as asigned
